refactor(scroller): extracts rendering logic for link rendering
